requirements.txt
langchain
python-dotenv
gunicorn
uvicorn
openai
faiss-cpu
unstructured
unstructured[local-inference]
tiktoken